requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Data Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# Image Processing & Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Annotation Tools
roboflow>=1.0.0
labelimg>=1.8.6

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter & Development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
click>=8.1.0
argparse>=1.4.0

# Optional: Web Interface
streamlit>=1.25.0
gradio>=3.35.0

# Optional: Model Serving
fastapi>=0.100.0
uvicorn>=0.23.0

# Development & Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
